moved ObservationMetaData and spatialBound to sims_utils
