+++ python/lsst/sims/catalogs/generation/db/__init__.py
@@ -1,4 +1,2 @@
-from .spatialBounds import *
-from .ObservationMetaData import *
 from .dbConnection import *
 from .utils import *
